# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP-сервер на Go с автоматической генерацией Swagger при старте",
        "title": "Configurator",
        "contact": {
            "name": "Lorgar Avrelian",
            "url": "https://github.com/Lorgar-Avrelian",
            "email": "victor-14-244@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://apache.org"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
...
        "/api/v1/oids/exact": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Парсер: OID"
                ],
                "summary": "Поиск OID по dotter notation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Точная dotter_notation",
                        "name": "notation",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Oid"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/oids/exact-with-mib": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Парсер: OID"
                ],
                "summary": "Поиск OID по dotter notation и названию MIB",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Точная dotter_notation",
                        "name": "notation",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Точное название MIB",
                        "name": "mib",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Oid"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/oids/exact-with-mib-and-vendor": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Парсер: OID"
                ],
                "summary": "Поиск OID по dotter notation, названию MIB и производителю",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Точная dotter_notation",
                        "name": "notation",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Точное название MIB",
                        "name": "mib",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Имя или директория вендора (передайте null или оставьте пустым для фильтрации по NULL вендорам)",
                        "name": "vendor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Oid"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/oids/mib": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Парсер: OID"
                ],
                "summary": "Получить OID по названию MIB",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название MIB",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Oid"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/oids/prefix": {
            "get": {
                "description": "Возвращает отсортированный по dotter notation список OID (по 100 на страницу)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Парсер: OID"
                ],
                "summary": "Поиск OID по префиксу с пагинацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Префикс dotter_notation",
                        "name": "prefix",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы (дефолт: 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.OidPageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/oids/vendor": {
            "get": {
                "description": "Находит производителя в кэше памяти, выгружает OID, сортирует и отдаёт по 100 штук на страницу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Парсер: OID"
                ],
                "summary": "Получить OID по производителю с пагинацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя вендора или его директория",
                        "name": "identity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы (дефолт: 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.OidPageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "configurator_internal_dto.OidPageResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configurator_internal_model.Oid"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "example": 100
                },
                "total_items": {
                    "type": "integer",
                    "example": 450
                }
            }
        },
...
        "configurator_internal_model.Oid": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string",
                    "example": "read-only"
                },
                "category": {
                    "type": "string",
                    "example": "system"
                },
                "description": {
                    "type": "string",
                    "example": "A textual description of the entity."
                },
                "dotter_notation": {
                    "type": "string",
                    "example": "1.3.6.1.2.1.1.1"
                },
                "enum": {
                    "type": "object"
                },
                "id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "mib_id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "sysDescr"
                },
                "number": {
                    "type": "integer",
                    "example": 1
                },
                "object_descriptor": {
                    "type": "string",
                    "example": "1.3.6.1.2.1.1.1"
                },
                "status": {
                    "type": "string",
                    "example": "current"
                },
                "syntax": {
                    "type": "string",
                    "example": "DisplayString"
                },
                "type": {
                    "type": "string",
                    "example": "OBJECT-TYPE"
                },
                "units": {
                    "type": "string",
                    "example": "seconds"
                }
            }
        },
...
    },
...
}